An efficient method to check if a number is multiple of 3
There is a pattern in binary representation of the number that
can be used to find if number is a multiple of 3. If difference
between count of odd set bits (Bits set at odd positions) and
even set bits is multiple of 3 then is the number.

Time Complexity is O(log n)

GFG Link :
https://www.geeksforgeeks.org/write-an-efficient-method-to-check-if-a-number-is-multiple-of-3/